add option to do intense HTTP logging
